crossplane-functions/users/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
idna==3.4
pathspec==0.10.3
platformdirs==2.6.2
PyYAML==6.0
tomli==2.0.1
urllib3==1.26.14
argparse==1.4.0